requirements.txt
Django
djangorestframework
PyYAML
psycopg2  # For PostgreSQL
gunicorn  # For production deployment
reportlab
paramiko
